chore(stack): change kafka's zookeeper.connect property
The last change added the docker-compose file and,
therefore, changed the way containers are run.
Before, they were run directly by the "docker run"
command, and the "network" property set to "host."
In this way, containers were able to communicate
over the ports with the host defined as
"localhost" by default.

Since the docker-compose runs them on a separate
docker network, Kafka no longer finds the
Zookeeper only by the port. Therefore, it was
necessary to adjust the Kafka's configuration and
set "zookeeper.connect" correctly.